+++ src/Exceptions/InvalidTemplatePathException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Kent013\PrismPrompt\Exceptions;
-
-use RuntimeException;
-
-final class InvalidTemplatePathException extends RuntimeException {}